:recycle: move test to SwdifffTests file
